Split the shared email module along the direction of travel
`email_common.py` was 601 lines and the name was the tell: it had become
wherever an email helper went. It held four unrelated jobs -- parsing a sender,
turning inbound markup into text, turning outbound content into markup, and
resolving attachments -- so every email provider imported the whole thing to
use a third of it.

They are four modules now, named for what they do:

  email_identity.py     who sent it, and which thread it belongs to
  email_text.py         inbound: HTML and quoted history to the words typed
  email_render.py       outbound: text/markdown/HTML to the parts a client shows
  email_attachments.py  paths to inline bytes or signed links

The seam that decides the split is direction. `email_text` reads what arrived
and `email_render` writes what leaves; they are mirrors, and keeping them in one
file hid that they never call each other.

All 20 import sites updated to the new homes -- no re-export facade, because a
module that exists only to forward names is the same "wherever it goes" problem
with an extra hop.

Also here, found while moving code past it:

- `teams/adapter.py` had `if email: pass` before its return. A branch that does
  nothing, counted as complexity, read as unfinished.
- The `payload_any` imports from the previous commit had been inserted above the
  stdlib block in five files. Ruff never said so: this repo sets no `select`, so
  it lints the default E4/E7/E9/F and isort is off entirely.
- `outlook/parser.py` imported `domain.entities` twice, one name each.
- The log event moved with its code: `agent_surfaces.email_common.*` is now
  `agent_surfaces.email_attachments.*`, catalog regenerated (651 events).

Baseline down: agent_surfaces complexity max 39 -> 33, total 935 -> 865,
functions over threshold 43 -> 41.

810 surfaces tests and 5,328 across the backend pass unchanged.
