+++ lemma-backend/app/core/log/event_catalog.py
@@ -123,7 +123,7 @@ EVENT_CATALOG: dict[str, EventSpec] = {
     'agent_surfaces.client.teams_token_acquisition_tenant_s.diagnostic': EventSpec('debug', frozenset({'error_code', 'status', 'tenant_id'})),
     'agent_surfaces.credential_resolver.could_not_refresh_credentials_account.diagnostic': EventSpec('debug', frozenset({'account_id'})),
     'agent_surfaces.credential_resolver.could_not_resolve_provider_account.diagnostic': EventSpec('debug', frozenset()),
-    'agent_surfaces.email_common.skipping_oversize_workspace_email_attachment.diagnostic': EventSpec('debug', frozenset({'count', 'inline_cap_bytes'})),
+    'agent_surfaces.email_attachments.skipping_oversize_workspace_email_attachment.diagnostic': EventSpec('debug', frozenset({'count', 'inline_cap_bytes'})),
     'agent_surfaces.email_surface_provisioning.address_unavailable.degraded': EventSpec('warning', frozenset({'agent_id', 'pod_id'})),
     'agent_surfaces.email_surface_provisioning.failed.degraded': EventSpec('warning', frozenset({'agent_id', 'failure_code', 'failure_type', 'pod_id'})),
     'agent_surfaces.event_receiver_service.could_not_load_telegram_polling.observed': EventSpec('debug', frozenset()),
