+++ lemma-backend/architecture-baseline.json
@@ -44,9 +44,9 @@
     "agent:count": 15,
     "agent:max": 42,
     "agent:total": 371,
-    "agent_surfaces:count": 43,
-    "agent_surfaces:max": 39,
-    "agent_surfaces:total": 935,
+    "agent_surfaces:count": 41,
+    "agent_surfaces:max": 33,
+    "agent_surfaces:total": 865,
     "connectors:count": 8,
     "connectors:max": 24,
     "connectors:total": 152,
@@ -95,8 +95,8 @@
     "app/modules/agent_surfaces/api/controllers/surface_controller.py": 627,
     "app/modules/agent_surfaces/api/schemas.py": 610,
     "app/modules/agent_surfaces/platforms/slack/service.py": 956,
-    "app/modules/agent_surfaces/platforms/teams/adapter.py": 787,
-    "app/modules/agent_surfaces/platforms/teams/service.py": 618,
+    "app/modules/agent_surfaces/platforms/teams/adapter.py": 786,
+    "app/modules/agent_surfaces/platforms/teams/service.py": 616,
     "app/modules/agent_surfaces/platforms/whatsapp/service.py": 623,
     "app/modules/agent_surfaces/services/event_receiver_service.py": 669,
     "app/modules/agent_surfaces/services/surface_service.py": 954,
